requirements.txt
fastapi
uvicorn
langchain
openai
pinecone-client
pypdf
python-dotenv

